Project_back/requirements.txt
appdirs==1.4.3
asgiref==3.2.6
astroid==2.3.3
b==1.0.0
colorama==0.4.3
distlib==0.3.0
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.2
pandas==1.0.3
patsy==0.5.1
psycopg2==2.8.5
pygame==1.9.6
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
statsmodels==0.11.1
virtualenv==20.0.10
wrapt==1.11.2
